Moved url functions into internal/url
Also merged the test cases.
