+++ database/util_test.go
@@ -1,7 +1,6 @@
 package database
 
 import (
-	"errors"
 	"testing"
 )
 
@@ -42,66 +41,6 @@ func TestGenerateAdvisoryLockId(t *testing.T) {
 				if tc.expectedID != "" {
 					t.Error("Got unexpected error:", err)
 				}
-			}
-		})
-	}
-}
-
-func TestSchemeFromUrlSuccess(t *testing.T) {
-	cases := []struct {
-		name     string
-		urlStr   string
-		expected string
-	}{
-		{
-			name:     "Simple",
-			urlStr:   "protocol://path",
-			expected: "protocol",
-		},
-		{
-			// See issue #264
-			name:     "MySQLWithPort",
-			urlStr:   "mysql://user:pass@tcp(host:1337)/db",
-			expected: "mysql",
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			u, err := SchemeFromURL(tc.urlStr)
-			if err != nil {
-				t.Fatalf("expected no error, but received %q", err)
-			}
-			if u != tc.expected {
-				t.Fatalf("expected %q, but received %q", tc.expected, u)
-			}
-		})
-	}
-}
-
-func TestSchemeFromUrlFailure(t *testing.T) {
-	cases := []struct {
-		name      string
-		urlStr    string
-		expectErr error
-	}{
-		{
-			name:      "Empty",
-			urlStr:    "",
-			expectErr: errors.New("URL cannot be empty"),
-		},
-		{
-			name:      "NoScheme",
-			urlStr:    "hello",
-			expectErr: errors.New("no scheme"),
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			_, err := SchemeFromURL(tc.urlStr)
-			if err.Error() != tc.expectErr.Error() {
-				t.Fatalf("expected %q, but received %q", tc.expectErr, err)
 			}
 		})
 	}
